Minor fixes

+++ src/Player.cpp
@@ -39,18 +39,19 @@ Player::~Player()
 void Player::openMedia(QString recordpath)
 {
     QString path = "ftp://support:Lvp^^ej0@192.168.0.8/ftp_media/usb/" + recordpath.remove(0, 16);
+
     player->setMedia(QUrl(path));
     player->play();
 }
 
-void Player::durationChanged(qint64 duration)
+void Player::durationChanged(qint32 duration)
 {
-    this->duration = duration/1000;
+    this->duration = duration / 1000;
 }
 
 void Player::updateTime()
 {
-    ui->duration->setText(QDateTime::fromTime_t(player->position()/1000).toUTC().toString("mm:ss") + " / " + QDateTime::fromTime_t(duration).toUTC().toString("mm:ss"));
+    ui->duration->setText(QDateTime::fromTime_t(player->position() / 1000).toUTC().toString("mm:ss") + " / " + QDateTime::fromTime_t(duration).toUTC().toString("mm:ss"));
 }
 
 void Player::stateChanged()
@@ -70,6 +71,7 @@ void Player::on_play_clicked()
     else
     {
         player->play();
+
         ui->play->setIcon(QIcon(":/images/pause.png"));
     }
 }
@@ -79,31 +81,36 @@ void Player::on_mute_clicked()
     if (player->isMuted())
     {
         player->setMuted(false);
+
         ui->mute->setIcon(QIcon(":/images/volume.png"));
     }
     else
     {
         player->setMuted(true);
+
         ui->mute->setIcon(QIcon(":/images/mute.png"));
     }
 }
 
 bool Player::eventFilter(QObject* watched, QEvent* event)
 {
-    if (watched == ui->slider && event->type() == QEvent::MouseButtonRelease )
+    if (watched == ui->slider && event->type() == QEvent::MouseButtonRelease)
     {
-        QMouseEvent *mouseEvent = static_cast<QMouseEvent *>(event);
+        QMouseEvent* mouseEvent = static_cast<QMouseEvent*>(event);
+
         player->setPosition(QStyle::sliderValueFromPosition(ui->slider->minimum(), ui->slider->maximum(), mouseEvent->x(), ui->slider->width()));
     }
 
-    if (watched == ui->volume && event->type() == QEvent::MouseButtonRelease )
+    if (watched == ui->volume && event->type() == QEvent::MouseButtonRelease)
     {
-        QMouseEvent *mouseEvent = static_cast<QMouseEvent *>(event);
+        QMouseEvent* mouseEvent = static_cast<QMouseEvent*>(event);
+
         player->setVolume(QStyle::sliderValueFromPosition(ui->volume->minimum(), ui->volume->maximum(), mouseEvent->x(), ui->volume->width()));
 
         if (player->isMuted())
         {
             player->setMuted(false);
+
             ui->mute->setIcon(QIcon(":/images/volume.png"));
         }
     }
